refactor: split Potassium command adapter by domain
Replace try/fail dispatch chains with focused session, combat, and boss
modules so the adapter stays a thin facade without changing scene behavior.
